Add option to seed refreshed test database
